validation: store type of sample_index and sample_mask builtins must be 'u32'
Bug: tint:506
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/55121
Kokoro: Kokoro <noreply+kokoro@google.com>
Auto-Submit: Sarah Mashayekhi <sarahmashay@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>

+++ src/resolver/entry_point_validation_test.cc
@@ -370,7 +370,7 @@ TEST_F(ResolverEntryPointValidationTest,
   // fn main(param : Input) {}
   auto* input = Structure(
       "Input",
-      {Member("a", ty.f32(),
+      {Member("a", ty.u32(),
               {Location(Source{{13, 43}}, 0),
                Builtin(Source{{14, 52}}, ast::Builtin::kSampleIndex)})});
   auto* param = Param("param", ty.Of(input));
